Continued WIP on -monom.  Separated fgg_monom.go.
Minor tidying.

Minor.

[WIP] Continued on -monom.

[WIP] Continued on -monom.

[WIP] Continued on -monom.

[WIP] Continued on -monom.

[WIP] Continued on -monom.
